Add files via upload
Air is what keeps humans alive. Monitoring it and understanding its quality is of immense importance to our well-being. For analyzing Air quality in India, a dataset was collected from the Database, this dataset provides information about different particles such as PM2.5, PM10, NO2, SO2 etc. which will affect human beings, animals and birds. examined the effects of particles in each city by using statistical operations on the given dataset. The findings explain which cities are less affected and which are more affected, as well as which cities are good from pollutant particles. 

Also calculated the percentage of AQI for different levels, which is very important to take control from it. Using the Pandas, Matplotlib libraries, visualized the statistical results to make them easier to understand. 